Move diagnostics to top level module
